Fix: Fix update comment error
- 코멘트 수정 시, 코멘트 생성 창에도 코멘트 내용이 세팅 되는 문제 수정
